testando configuração de execução de testes automatizados no github actions

+++ src/test/java/AccessFileParserTest.java
@@ -187,16 +187,19 @@ public class AccessFileParserTest extends ParserBaseTest {
     @AfterClass
     public static void tearDown()
     {
-        try 
+        if(!isRunningOnGithub)
         {
-         for (var format : fileformats) 
+            try 
             {
-                var fileName = "splinker_%s.db".formatted(format.name());
-                Files.delete(Path.of(fileName));            
-            }   
-        } catch (Exception e) 
-        {
-            e.printStackTrace();
+            for (var format : fileformats) 
+                {
+                    var fileName = "splinker_%s.db".formatted(format.name());
+                    Files.delete(Path.of(fileName));            
+                }   
+            } catch (Exception e) 
+            {
+                e.printStackTrace();
+            }
         }
     }
